% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc-test.R
\name{assoc}
\alias{assoc}
\title{title.}
\usage{
assoc(data, re = NULL)
}
\arguments{
\item{par.}{}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
details.
}
\author{
Erick Albacharro Chacon-Montalvan
}
