\name{BeginTable}
\alias{BeginTable}
\title{LaTeX: Begin Table}
\usage{BeginTable(loc, ...)
}
\description{Print to console: begin table environment.}
\value{Nothing.}
\author{Vinh Nguyen}
\arguments{\item{loc}{Location of table, e.g, \code{htb}; defaults to nothing.}
\item{...}{Nothing.}
}
\examples{BeginTable() ; BeginTable("htb")}

