\name{BeginTabular}
\alias{BeginTabular}
\title{LaTeX: Begin Tabular}
\usage{BeginTabular(align, hline.num=1, ...)
}
\description{Print to console: begin tabular environment.}
\value{Nothing.}
\author{Vinh Nguyen}
\arguments{\item{align}{Tabular alignment, e.g., \code{"lrr"}}
\item{hline.num}{Number of horizontal lines following tabular statement; defaults to 1.}
\item{...}{Nothing.}
}
\examples{BeginTabular("lrr"); BeginTabular("lrr", 2)}

