% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{condense2}
\alias{condense2}
\title{Efficient 2d binnary summaries.}
\usage{
condense2(x, y, var = NULL, data, width, origin, ...)
}
\arguments{
\item{x}{the first variable on data to be binned.}

\item{y}{the second variable on date to be binned.}

\item{var}{varible to be summarised on x, y bins.}

\item{data}{data.frame where the variables exist.}

\item{width}{vector of bin widths for \code{\link{bin}} function.}

\item{origin}{vector for \code{\link{bin}} function.}

\item{...}{additional options passed to \code{\link{condense}} function.}
}
\value{
condense object.
}
\description{
\code{condense2} creates a two dimensional binnary summary with bigvis
package.
}
\details{
\code{condense2} is compatible with dplyr.
}
\author{
Erick A. Chacon-Montalvan
}

