% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor-gen.R
\name{corr}
\alias{corr}
\title{title.}
\usage{
corr(data)
}
\arguments{
\item{par.}{}
}
\value{
return.
}
\description{
\code{function} description.
}
\details{
details.
}
\author{
Erick Albacharro Chacon-Montalvan
}

