% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-handle.R
\name{cumsum_na}
\alias{cumsum_na}
\title{Cumsum vector with NA values.}
\usage{
cumsum_na(x, ...)
}
\arguments{
\item{x}{}
}
\description{
\code{cumsum_na} Cumsum vector with NA values.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}

