% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-custom.R
\name{cut_ext}
\alias{cut_ext}
\title{Convert Numeric to Factor with custom condition}
\usage{
cut_ext(x, breaks, condition = NULL)
}
\arguments{
\item{x}{numeric vector to be converted to factor.
breaks numeric vector of breaks to be included.
condition condition to be applied to the high threshold of interval.}
}
\value{
The new factor is returned.
}
\description{
\code{cut_ext} converts a numeric object to factor given the breaks and
condition applied to the high threshold of each interval.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}

