% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day2day.R
\docType{package}
\name{day2day}
\alias{day2day}
\alias{day2day-package}
\title{\code{day2day} package}
\description{
Functions that can be useful in the day-to-day data analysis. It comprehends
functions to find paths for projects, make summaries of databases inside
folder and so on.
}
