% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-summarize.R
\name{db_summarize}
\alias{db_summarize}
\title{Summarize the databases inside a folder.}
\usage{
db_summarize(path_data, filename = "summary-databases.txt")
}
\arguments{
\item{file}{file name to look for.}
}
\value{
A character indicating the folder path.
}
\description{
\code{db_summary} write the summary of databases, founded inside a provided
folder, into a file.
}
\details{
When the filename is not specified, \code{db_summary} will create the file
"summary-databases.txt".
}
\author{
Erick Albacharro Chacon-Montalvan
}

