% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical.R
\name{ecdf_val}
\alias{ecdf_val}
\title{Compute empirical cumulative density function.}
\usage{
ecdf_val(x)
}
\arguments{
\item{x}{Vector of realizations.}
}
\value{
vector of quantiles.
}
\description{
\code{ecdf_val} Computes the empirical quantile using \code{ecdf} funcion.
}
\details{
details.
}
\examples{

x <- rnorm(1000)
ecdf_x <- ecdf_val(x)
plot(x, ecdf_x)

}
\author{
Erick A. Chacon-Montalvan
}
