\name{EndTabular}
\alias{EndTabular}
\title{LaTeX: End Tabular}
\usage{EndTabular(hline.num=1, ...)
}
\description{Print to console: end tabular environment.}
\value{Nothing.}
\author{Vinh Nguyen}
\arguments{\item{hline.num}{Number of horizontal lines before tabular statement; defaults to 1.}
\item{...}{Nothing.}
}
\examples{EndTabular(1); EndTabular(2)}

