% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spi-rain.R
\name{find_flood_drought}
\alias{find_flood_drought}
\title{Identify floods and droughts based on SPI.}
\usage{
find_flood_drought(spi)
}
\arguments{
\item{par.}{}
}
\value{
return.
}
\description{
\code{find_flood_drought} description.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}
