% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path.R
\name{getwd_path}
\alias{getwd_path}
\title{Cut Current Working Directory Filepath}
\usage{
getwd_path(foldername)
}
\arguments{
\item{foldername}{Folder name that contains current working directory.}
}
\value{
A character indicating the folder path.
}
\description{
\code{getwd_path} finds the absolute filepath of a folder that contains the
current working directory.
}
\details{
\code{getwd_path} gets the filepath of the current working directory and shortens
it given a folder name inside this path.
}
\author{
Erick A. Chacon-Montalvan
}
