% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path.R
\name{getwd_short}
\alias{getwd_short}
\title{Cut Current Working Directory Filepath}
\usage{
getwd_short(foldername)
}
\arguments{
\item{foldername}{Folder name that contains current working directory.}
}
\value{
A character indicating the folder path.
}
\description{
\code{getwd_short} finds the absolute filepath of a folder that contains the
current working directory.
}
\details{
\code{getwd_short} gets the filepath of the current working directory and shortens
it given a folder name inside this path.
}
\examples{

\dontrun{
# Set working directory to tempdir
project_path <- file.path(tempdir(), "my-project")
dir.create(project_path)
code_path <- file.path(project_path, "analysis")
dir.create(code_path)
setwd(code_path)

# Find the path of the file
getwd()
getwd_short(foldername = "my-project")
}

}
\author{
Erick A. Chacon-Montalvan
}
