% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{ggscale_seq}
\alias{ggscale_seq}
\title{Creates sequence for continuous scale .}
\usage{
ggscale_seq(by = 1)
}
\arguments{
\item{by}{desired step of the sequence as in \code{\link{seq}}.}
}
\value{
a function that can be provided to \code{\link{scale_x_continuos}}
or similar ones.
}
\description{
\code{ggscale_seq} Creates a function that can be provide to compute
break for a continuous scale axis.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}

