% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{ggtheme}
\alias{ggtheme}
\title{Customized theme for ggplot.}
\usage{
ggtheme(color = "#990000")
}
\arguments{
\item{color}{Desired text color for the theme.}
}
\value{
theme for being used with ggplot objects.
}
\description{
\code{ggtheme} creates a customized theme for ggplot.
}
\details{
details.
}
\examples{

# Create a dataset
library(ggplot2)
data <- data.frame(x = rnorm(100))

# Plot variable x with custom theme
ggplot(data, aes(x)) +
  geom_histogram() +
  ggtheme()

}
\author{
Erick A. Chacon-Montalvan
}
