% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path.R
\name{git_path}
\alias{git_path}
\title{Get Path of Current Git Repository}
\usage{
git_path()
}
\value{
A character indicating the git repository path.
}
\description{
\code{git_path} finds the path of the current git repository.
}
\details{
Git is required to use \code{git_path} and it should be only used when working
under a git repository.
}
\author{
Erick A. Chacon-Montalvan
}
