\name{ifelse1}
\alias{ifelse1}
\title{ifelse1}
\usage{ifelse1(test, yes, no)
}
\description{Return \code{yes} if \code{test} is \code{TRUE}, else return \code{no}.
A shorthand way to do \link[base]{Control} flow.}
\value{\code{yes} or \code{no}.}
\arguments{\item{test}{An expression returning \code{\link[base]{logical}}.}
\item{yes}{Expression that is evaluated if \code{test} is \code{TRUE}.}
\item{no}{Expression that is evaluated if \code{test} is \code{FALSE}.}
}

