\name{is.whole}
\alias{is.whole}
\title{Whole Numbers}
\usage{is.whole(x)
}
\description{Determine whether each element of a vector or matrix of numerics is whole.}
\value{Vector or matrix containing \code{\link[base]{logical}} indicators.}
\author{Vinh Nguyen}
\arguments{\item{x}{Vector or matrix of numerics.}
}
\examples{is.whole(1:5) ; is.whole(1) ; is.whole(c(1,1.2))}

