\name{latex2html}
\alias{latex2html}
\title{LaTeX to html}
\usage{latex2html(file)
}
\description{Convert a LaTeX file to html.}
\details{Convert a \code{.tex} file to a \code{.html} file via \code{htlatex} from \code{tex4ht}.  This function will only work on *nix platforms.}
\value{Nothing.}
\author{Vinh Nguyen}
\references{http://www.tug.org/applications/tex4ht/mn.html
http://biostat.mc.vanderbilt.edu/wiki/Main/SweaveConvert}
\arguments{\item{file}{A file ending in \code{.tex}.}
}

