\name{LatexNewLine}
\alias{LatexNewLine}
\title{LaTeX: New Line}
\usage{LatexNewLine(n=1, ...)
}
\description{Print to console: \code{ \\\\} followed by a new line.}
\value{Nothing.}
\author{Vinh Nguyen}
\arguments{\item{n}{Number of new lines with \code{ \\\\}; defaults to 1.}
\item{...}{Nothing.}
}
\examples{LatexNewLine()}

