% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-info.R
\name{ls_size}
\alias{ls_size}
\title{Obtain list of objects by memory usage.}
\usage{
ls_size(units = "MB", ...)
}
\arguments{
\item{units}{Unit to be printed as with \code{object.size} function.}

\item{...}{Additional parameters passed to \code{ls}.}
}
\value{
vector of objects size sorted by memory usage.
}
\description{
\code{ls_size} obtain the list of objects with their memory usage.
}
\examples{

x <- c("1", "b")
ls_size(units = "KB")

}
\author{
Erick A. Chacon-Montalvan
}
