% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-handle.R
\name{mean_na}
\alias{mean_na}
\title{Sum vector with NA values.}
\usage{
mean_na(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{Parameters passed to \code{mean}}
}
\description{
\code{mean_na} Mean vector with NA values.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}
