% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{mgp}
\alias{mgp}
\title{Simulate a Multivariate Gaussian process}
\usage{
mgp(s1, s2, cov.model = NULL, variance = NULL, nugget = NULL,
  phi = NULL, kappa = NULL)
}
\arguments{
\item{s1}{First coordinate}

\item{s2}{Second coordinate}

\item{cov.model}{A character or function indicating the covariance function that
Should be used to compute the variance-covariance matrix}

\item{variance}{A qxq matrix of non-spatial covariance.}

\item{nugget}{A qxq diagonal matrix of non-spatial noise.}

\item{phi}{A q-length vector of decay parameters.}

\item{kappa}{A q-length vector of kappa parameters if Matern spatial correlation
function is used.}
}
\value{
A vector of the realization of the Gaussian Process
}
\description{
\code{mgp} Simulate a Multivariate spatial Gaussian process known as linear model
of coregionalization Y(h) = AS(h), where S(h) is a vector of q independent Gaussian
processes.
}
\details{
details.
}
\examples{

# Generate coordinates
N <- 100
s1 <- 2 * runif(N)
s2 <- 2 * runif(N)

# Covariance parameters
q <- 2
var <- sqrt(diag(c(4, 4)))
A <- matrix(c(1, - 0.8, 0, 0.6), nrow = 2)
variance <- var \%*\% tcrossprod(A) \%*\% var
nugget <- diag(0, q)
phi <- rep(1 / 0.08, q)

# Generate the multivariate Gaussian process
y <- mgp(s1, s2, "exponential", variance, nugget, phi)
y1 <- y[1:N]
y2 <- y[(N + 1):(2 * N)]

# Check correlation
cor(y1, y2)
plot(y1, y2)

# Visualize the spatial
plot(s1, s2, cex = y1, col = 2)
points(s1, s2, cex = y2, col = 3)

}
\author{
Erick A. Chacon-Montalvan
}
