% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{msim_model}
\alias{msim_model}
\title{Simulate Databases Based in Multivariate Models}
\usage{
msim_model(formula, generator = rnorm, n = 100, init_data = NULL,
  seed = NULL, extent = 1)
}
\arguments{
\item{formula}{List of the parameters, indicating how they should be computed.
similar to formula for \code{lm}, \code{glm}, \code{bamlss}, with the difference
that it included the coefficients and link function explicitly.}

\item{generator}{Function to generate the response variables given the parameters}

\item{n}{Number of observations to be simulated}

\item{init_data}{Initial data including some variables to not been simulated.}

\item{seed}{Seed to be defined with function \code{set.seed} to obtain reproducible
results}

\item{extent}{Spatial extent for the simulation of coordinates when a spatial effect
is included.}
}
\value{
a \code{tibble} containing the simulated predictors, parameters and response
variable
}
\description{
\code{msim_model} simulate a database based on common multivariate models. The
structure used to create the data is similar as the \code{bamlss.formula}.
}
\examples{

# Covariance parameters
n <- 100
q <- 2
var <- sqrt(diag(c(4, 4)))
A <- matrix(c(1, - 0.8, 0, 0.6), nrow = 2)
variance <- var \%*\% tcrossprod(A) \%*\% var
nugget <- diag(0, q)
phi <- rep(1 / 0.08, q)

# Structure of the model
formula <- list(
  mean ~ psych::logistic(
    mgp(s1, s2, "exponential", get("variance"), get("nugget"), get("phi"))),
  sd ~ 1
)

# Simulate data based on formula
library(tidyr)
library(dplyr)
data <- msim_model(formula, generator = rnorm, n = n, extent = 2, seed = 1)
data_long <- gather(data, yname, yval, matches("^y[0-9]+"))

# Plot the observed realization
library(ggplot2)
spgg <- ggplot(data_long, aes(s1, s2, size = yval, col = yval)) +
  geom_point() +
  scale_colour_gradientn(colours = terrain.colors(10)) +
  facet_wrap(~ yname)
print(spgg)


}
\author{
Erick Albacharro Chacon-Montalvan
}
