% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jekyll.R
\name{post_jekyll}
\alias{post_jekyll}
\title{Prepare md post to jekyll.}
\usage{
post_jekyll(filename,
  path_jekyll = "/home/chaconmo/Documents/Repositories/erickchacon/_posts/",
  fig_jekyll = "{{site.baseurl}}/assets/images/")
}
\arguments{
\item{filename}{Name of the post you want to export to jekyll.}

\item{path_jekyll}{Path to your _posts folder.}

\item{fig_jekyll}{String that needs to be added to the figures link.}
}
\value{
Generates a markdown file on your _posts folder (you must specify
this path adequately).
}
\description{
\code{post_jekyll} modify the markdown file created with \code{knit} and
exports it to the folder of your jekyll webpage. Basically, this function
modifies the path of images to match adequately with your jekyll structure
end save it on the _posts folder of your jekyll web repositoory.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}

