% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runmean}
\alias{runmean}
\title{Moving average.}
\usage{
runmean(a, width)
}
\arguments{
\item{a}{Vector of values representing the original time series.}

\item{width}{The number of values to consider for the average.}
}
\value{
A moving average of the time series.
}
\description{
\code{runmean} computes the moving average of a time series.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}

