% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{stat_events}
\alias{stat_events}
\title{Extreme events.}
\usage{
stat_events(mapping = NULL, data = NULL, geom = "ribbon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping created by \code{aes} or \code{aes}.}

\item{data}{Dataset to use.}

\item{geom}{Geometry.}

\item{position}{Position.}

\item{na.rm}{Logical value to remove missing values.}

\item{show.legend}{Show legend.}

\item{inherit.aes}{Inherit aesthetics.}

\item{...}{Additional arguments passed to \code{layer}.}
}
\description{
\code{stat_events} computes a data frame to adequatly draw areas
corresponding to extreme events in a time series.
}
\author{
Erick A. Chacon-Montalvan
}
