% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-handle.R
\name{sum_na}
\alias{sum_na}
\title{Sum vector with NA values.}
\usage{
sum_na(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{Parameters passed to \code{sum}}
}
\description{
\code{sum_na} Sum vector with NA values.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}

