% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{wrapit}
\alias{wrapit}
\title{Wrap text for ggplots.}
\usage{
wrapit(text, width = 20)
}
\arguments{
\item{text}{long text to be wrapped}

\item{width}{width of the wrap}
}
\value{
wrapped text
}
\description{
\code{wrapit} wrap long texts to be used with ggplot package.
}
\examples{

long_text <- paste0(rep("this is a very long text", 10), collapse = " ")
cat(long_text)
cat(wrapit(text = long_text, 26))

}
\author{
Erick A. Chacon-Montalvan
}
