% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc-test.R
\name{assoc}
\alias{assoc}
\title{Matrix of General Association}
\usage{
assoc(data, re = NULL)
}
\arguments{
\item{data}{Data frame with covariates}

\item{re}{Logical values, indicating if random effects should be included or not.}
}
\value{
A matrix of the pairwise association.
}
\description{
\code{assoc} computes a matrix of general association between categorical and
quantitative variables.
}
\details{
Given a \code{data.frame}, it will compute a matrix of association between all the
variables.
}
\author{
Erick Albacharro Chacon-Montalvan
}
