% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-custom.R
\name{cut_ext}
\alias{cut_ext}
\title{Convert Numeric to Factor with custom condition}
\usage{
cut_ext(x, breaks, condition = NULL)
}
\arguments{
\item{x}{numeric vector to be converted to factor.
breaks numeric vector of breaks to be included.
condition condition to be applied to the high threshold of interval.}

\item{breaks}{break point to divide the categories.}

\item{condition}{condition for each break point in order to define open or
close intervals.}
}
\value{
The new factor is returned.
}
\description{
\code{cut_ext} converts a numeric object to factor given the breaks and
condition applied to the high threshold of each interval.
}
\details{
details.
}
\examples{
cut_ext(1:10, breaks = c(3, 6), condition = c('<', '<='))

}
\author{
Erick A. Chacon-Montalvan
}
