% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-summarize.R
\name{db_summarize}
\alias{db_summarize}
\title{Summarize the databases inside a folder.}
\usage{
db_summarize(path_data, filename = "summary-databases.txt")
}
\arguments{
\item{path_data}{The path where to look for .Rdata files.}

\item{filename}{File name where to write the summary.}
}
\value{
A character indicating the folder path.
}
\description{
\code{db_summary} write the summary of databases, founded inside a provided
folder, into a file.
}
\details{
When the filename is not specified, \code{db_summary} will create the file
"summary-databases.txt".
}
\examples{

# Create a path with datasets to apply our function
data_path <- tempdir()
save(iris, file = file.path(data_path, "iris.RData"))
save(airquality, file = file.path(data_path, "airquality.RData"))

# Obtain the summary of all datasets included in an specific path
db_summarize(data_path, filename = file.path(data_path, "summary.txt"))

# Print the summary
readLines(file.path(data_path, "summary.txt"))

}
\author{
Erick Albacharro Chacon-Montalvan
}
