% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path.R
\name{get_path}
\alias{get_path}
\title{Get the Folder Path Containing File}
\usage{
get_path(file = "repos.ecm")
}
\arguments{
\item{file}{File name to look for.}
}
\value{
A character indicating the folder path.
}
\description{
\code{get_path} returns the folder path that contains a specified file.
}
\details{
When the file name is not specified, \code{get_path} will look for the
file "doctoral-research.ecm".
}
\author{
Erick Albacharro Chacon-Montalvan
}
