% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{ggtheme}
\alias{ggtheme}
\title{Customized theme for ggplot.}
\usage{
ggtheme(color = "#990000")
}
\arguments{
\item{color}{}
}
\value{
theme for being used with ggplot objects.
}
\description{
\code{ggtheme} creates a customized theme for ggplot.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}
