% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{mgp}
\alias{mgp}
\title{Simulate a Multivariate Gaussian process}
\usage{
mgp(s1, s2, cov.model = NULL, variance = NULL, nugget = NULL,
  phi = NULL, kappa = NULL)
}
\arguments{
\item{s1}{First coordinate}

\item{s2}{Second coordinate}

\item{cov.model}{A character or function indicating the covariance function that
Should be used to compute the variance-covariance matrix}

\item{variance}{A qxq matrix of non-spatial covariance.}

\item{nugget}{A qxq diagonal matrix of non-spatial noise.}

\item{phi}{A q-length vector of decay parameters.}

\item{kappa}{A q-length vector of kappa parameters if Matern spatial correlation
function is used.}
}
\value{
A vector of the realization of the Gaussian Process
}
\description{
\code{mgp} Simulate a Multivariate spatial Gaussian process known as linear model
of coregionalization Y(h) = AS(h), where S(h) is a vector of q independent Gaussian
processes.
}
\details{
details.
}
\examples{
# Generate coordinates
N <- 1000
s1 <- 2 * runif(N)
s2 <- 2 * runif(N)
# Simulate and plot the realization of a Gaussian process
y <- gp(s1, s2, "exp_cov", list(phi = 0.05, sigma2 = 1))
plot(s1, s2, cex = y)
# Plot with ggplot
# ggplot(data.frame(s1, s2, y), aes(s1, s2, col = y)) +
#  geom_point(size = 3)

}
\author{
Erick A. Chacon-Montalvan
}
