% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{msim_model}
\alias{msim_model}
\title{Simulate Databases Based in Multivariate Models}
\usage{
msim_model(formula, generator = rnorm, n = 100, init_data = NULL,
  seed = NULL, extent = 1)
}
\arguments{
\item{formula}{List of the parameters, indicating how they should be computed.
similar to formula for \code{lm}, \code{glm}, \code{bamlss}, with the difference
that it included the coefficients and link function explicitly.}

\item{generator}{Function to generate the response variables given the parameters}

\item{n}{Number of observations to be simulated}

\item{init_data}{Initial data including some variables to not been simulated.}

\item{seed}{Seed to be defined with function \code{set.seed} to obtain reproducible
results}

\item{extent}{Spatial extent for the simulation of coordinates when a spatial effect
is included.}
}
\value{
a \code{tibble} containing the simulated predictors, parameters and response
variable
}
\description{
\code{msim_model} simulate a database based on common multivariate models. The
structure used to create the data is similar as the \code{bamlss.formula}.
}
\examples{
f <- list(
 mean ~ 5 + 0.5 * x1 + 0.1 * x2 + 0.7 * id1,
 sd ~ exp(x1)
)
data <- sim_model(f, rnorm, 100)

}
\author{
Erick Albacharro Chacon-Montalvan
}
