% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{stat_events}
\alias{stat_events}
\title{Extreme events.}
\usage{
stat_events(mapping = NULL, data = NULL, geom = "ribbon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Aesthetic mapping created by \code{aes} or \code{aes}.}

\item{threshold}{Threshold to plot the extreme events.}

\item{n}{Number of points used for interpolation.}
}
\description{
\code{stat_events} computes a data frame to adequatly draw areas
corresponding to extreme events in a time series.
}
\details{
details.
}
\author{
Erick A. Chacon-Montalvan
}
