% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-adds.R
\name{wrapit}
\alias{wrapit}
\title{Wrap text for ggplots.}
\usage{
wrapit(text, width = 20)
}
\arguments{
\item{text}{long text to be wrapped}

\item{width}{width of the wrap}
}
\value{
wrapped text
}
\description{
\code{wrapit} wrap long texts to be used with ggplot package.
}
\author{
Erick A. Chacon-Montalvan
}
